% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEdata.R
\name{list.names}
\alias{list.names}
\title{Get list names}
\usage{
list.names(mse_data)
}
\arguments{
\item{mse_data}{object of class `MSEdata`.}
}
\value{
names of the MSE lists.
}
\description{
Get list names
}
