% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEdata.R
\name{merge}
\alias{merge}
\title{Merge lists}
\usage{
merge(mse_data, ...)
}
\arguments{
\item{mse_data}{MSEdata object from which to merge lists.}

\item{...}{names of the lists to be merged.}
}
\description{
Merge lists
}
