% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMSE}
\alias{plotMSE}
\title{UpSetR plot of MSE data}
\usage{
plotMSE(
  mse_data,
  sets.bar.color = "cornflowerblue",
  main.bar.color = "cornflowerblue",
  matrix.color = "cornflowerblue",
  mainbar.y.label = "Count",
  sets.x.label = "",
  ...
)
}
\arguments{
\item{mse_data}{todo}

\item{sets.bar.color}{todo}

\item{main.bar.color}{todo}

\item{matrix.color}{todo}

\item{mainbar.y.label}{todo}

\item{sets.x.label}{todo}

\item{...}{todo}
}
\description{
UpSetR plot of MSE data
}
